Initial commit
This database is to be used in conjunction with the MSTFastDev.py development environment.  You will need to create a MySQL database named mst_fast before you import this dump file.  It contains data on 500 locations and the 124,750 distances between each distinct pairs of points.  A fast MST algorithm can compute an MST solution in a relatively few seconds.